test sign in, test edit user password errors, test edit user password_confirmation errors
